LayoutTests/platform/qt/fast/multicol/vertical-rl/break-properties-expected.txt

PASS: 'break-before' is at (748, 220)
PASS: 'after-break' is at (748, 431)
